chore(pano): restructure components in ui-library
