+++ apps/pano/src/entry.client.tsx
@@ -1,6 +1,6 @@
 import { RemixBrowser } from "@remix-run/react";
 import ReactDOM from "react-dom";
-import { ClientCacheProvider } from "~/ui-library/StyleProvider";
+import { ClientCacheProvider } from "~/ui-library";
 
 ReactDOM.hydrate(
   <ClientCacheProvider>
