+++ apps/pano/src/features/post/PostItem.tsx
@@ -2,16 +2,11 @@ import { useFetcher } from "@remix-run/react";
 import normalizeUrl from "normalize-url";
 import type { FC } from "react";
 import type { Post } from "~/models/post.server";
-import {
-  Box,
-  ExternalLink,
-  GappedBox,
-  SmallLink,
-  UpvoteButton,
-} from "~/ui-library";
-import { MoreOptionsDropdown } from "~/ui-library/MoreOptionsDropdown";
+import { Box, ExternalLink, GappedBox, SmallLink } from "~/ui-library";
 import { canUserEdit } from "../auth/can-user-edit";
 import { useUserContext } from "../auth/user-context";
+import { UpvoteButton } from "../upvote/UpvoteButton";
+import { MoreOptionsDropdown } from "./MoreOptionsDropdown";
 
 type PostItemProps = {
   post: Post;
