+++ apps/pano/src/features/topnav/SearchInput.tsx
@@ -1,11 +1,6 @@
 import { MagnifyingGlassIcon } from "@radix-ui/react-icons";
 import type { FC } from "react";
-import { ControlGroup } from "../ControlGroup";
-import { Form } from "../Form";
-import { Input } from "../Input";
-import { Select } from "../Select";
-import { Button } from "../layout-components/Button";
-
+import { Button, ControlGroup, Form, Input, Select } from "~/ui-library";
 const hasCollections = () => false;
 
 export const SearchInput: FC = () => {
