+++ apps/pano/src/features/upvote/UpvoteButton.tsx
