+++ apps/pano/src/features/user-dropdown/UserDropdown.tsx
@@ -2,16 +2,16 @@ import { useSubmit } from "@remix-run/react";
 import type { FC } from "react";
 import { styled } from "~/stitches.config";
 import {
+  Button,
   DropdownMenu,
   DropdownMenuArrow,
   DropdownMenuContent,
   DropdownMenuItem,
   DropdownMenuSeparator,
   DropdownMenuTrigger,
-} from "./Dropdown";
-import { Link } from "./Link";
-import { UserAvatar } from "./UserAvatar";
-import { Button } from "./layout-components/Button";
+  Link,
+  UserAvatar,
+} from "~/ui-library";
 
 const MenuItem = styled(DropdownMenuItem, {
   width: "auto",
